requirements.txt
matplotlib
numpy
pandas
pyzmq
spacy